requirements.txt
selenium
webdriver-manager
pandas
requests
beautifulsoup4
pymysql
apache-airflow-providers-mysql
apache-airflow[postgres]